refactor: utils.benchmarks module
